Moved the reporters into a folder, more debugging
